fixed flake8/pep8 violations, excluded videos from mock_data, redid tests, and moved batch.py to pytube/pytube/contrib
